Refactor ASR and translation services; add validation API and audio player hooks
